requirements.txt
pylint == 2.12.2
PyGithub == 1.55